Fix Name::root
The in-memory representation used was correct, but the size information
in the reference returned was incorrect, since it was computed from the
length of the entire structure (= the length of ROOT_NAME_REPR), not the
length of the dynamically sized portion. To fix this, the technique used
by Name::make_fat_pointer is employed. Accordingly, what was
Name::make_fat_pointer now comes in const (Name::make_fat_pointer) and
mut (Name::make_fat_pointer_mut) variants, and Name::root employs the
(new) const version.