Don't implement Eq or PartialEq for Rdata
The old implementations compared the underlying octet buffers bitwise.
This is not always the right thing to do: one might want a semantic
comparison, which cannot be expressed using the Eq/PartialEq traits.
(Such semantic comparison requires the caller to provide the RR type,
and it may fail on invalid inputs.)

To prevent confusion, this removes the Eq and PartialEq implementations
for Rdata, which forces the caller to choose explicitly the right
behavior:

* For bitwise comparison, use rdata1.octets() == rdata2.octets().
* For semantic comparison, use the Rdata::equals and Rdata::equals_as_*
  methods.