Make Rcode a struct like Class, Type and such
It's hard to hide the Rcode::Unassigned variant's data to prevent the
creation of "unassigned" RCODEs that are actually assigned. This problem
gets even worse when we start considering extended RCODEs, whose values
are almost all unassigned! Therefore, it seems best to make Rcode a
single-member tuple struct wrapping a (private) u8, and to plan to make
ExtendedRcode a single-member tuple struct wrapping a u16. Now they'll
be like Class, Type, and similar data types.