Add an independent size limit to message::Writer
A size limit less than the underlying buffer length can now be set when
a Writer is constructed. It can be increased later through the
Writer::increase_limit method.

The Writer structure now also keeps track of the "available" size. This
currently remains the same as the limit, but it will be used in the
future to reserve space for meta-RRs like OPT or TSIG that will be added
when the message is finished.