Allow the Catalog to be updated at any time
This wraps a Server's Catalog in an Arc protected by a RwLock.
Processing code obtains a read lock and clones the Arc to create a
snapshot of the Catalog when beginning to process each message.

It may be beneficial for performance eventually to use something like
the arc-swap crate, but (a) this works for now and (b) this requires
nothing but the standard library.