Add local address support on NetBSD
The API for IPv4 sockets on NetBSD is essentially the same as that of
Linux. The one difference is that struct in_pktinfo on NetBSD does not
have the ipi_spec_dst field; the ipi_addr field is used with both
recvmsg and sendmsg. (NetBSD's netinet/in.h actually has
"#define ipi_spec_dst ipi_addr" to maintain source compatibility with
Linux and Solaris, but Rust's libc crate understandably can't do that.)

NetBSD also has IP_RECVDSTADDR and IP_SENDSRCADDR control messages,
which provide the same functionality with the API used in FreeBSD and
OpenBSD. However, the IP_RECVPKTINFO/IP_PKTINFO-based API seems to have
come first, and in any case, the nix crate does not currently support
IP_SENDSRCADDR. So we'll stick with IP_RECVPKTINFO/IP_PKTINFO for
NetBSD.