Interpret zone file paths relative to config file
When loading the configuration from a file, zone file paths are now
interpreted relative to the configuration file. For example, if the
configuration file is /etc/quandary/config.toml and a zone is to be
loaded from example.zone, Quandary will try to open
/etc/quandary/example.zone.

When zones are specified directly on the command line, the previous
behavior (searching relative to the working directory) is retained.