Add a Dockerfile
Quandary will be built using the official Rust image (currently, the
Debian 11 version) and then will be installed into a Google distroless
image (currently, the non-root Debian 11 version with glibc and libgcc).
This gives a final image size of under 30 MB.

Quandary will run as a non-root user (65532 in the current distroless
image), but the binary is granted the CAP_NET_BIND_SERVICE capability so
that it may bind to port 53. The working directory is set to
/etc/quandary. By default, it will attempt to load its configuration
from /etc/quandary/config.toml.