Use larger UDP payload sizes with EDNS
The default EDNS UDP payload size is 1,232 octets, as recommended for
DNS Flag Day 2020.

With these changes, it also seemed wise to make Server::handle_message
require callers to pass a large-enough response_buf (65,535 octets for
TCP and the configured EDNS UDP payload size for UDP). Previously, it
would work with shorter buffers (possibly leading to SERVFAILs or
truncated responses). It will now panic if this requirement is not met.