Rework zone_file::Line
For upcoming $INCLUDE work, it seems like it will be best to keep the
line member out of ParsedRr and Include. The Line type is now a struct
that contains the line number itself, independent of the line's content
(Include or ParsedRr). For the file-based parser that has $INCLUDE
support, we'll use a counterpart Line struct that has both the line
number *and* the file path.