Rework the zone module into a new db module
The zone module focused on concrete types for storing and accessing DNS
zones in memory. The new db module, by contrast, turns the old Catalog
and Server structures into traits so that different in-memory
data structures and even other backends (e.g. SQL databases) might be
used for the server's database. The old Catalog and Zone implementations
remain as the ready-made HashMapTreeCatalog and HashMapTreeZone
structures.

The zone lookup API has undergone significant revision. A new
LookupOptions structure controls lookup behavior, including whether to
search below zone cuts, and is required on all lookup methods. This will
reduce the number of separate methods that do almost the exact same
thing (the direction that the previous API was going). Lookup result
types now use std::borrow::Cow to allow the on-the-fly construction of
RRsets in addition to in-memory storage. There are also now three types
of lookup:

* lookup of a specific RRset by type;
* lookup of address RRsets (A and AAAA); and
* lookup of all RRsets.

The new type (lookup of addresses) is now used for additional section
processing and glue record lookup (where a lookup of all RRsets was
previous used). Since the lookup of all records is now used only for *
(ANY) queries, it also makes it possible to eventually make this
functionality optional. The overall goal of the changes is to make the
implementation of non-memory backends feasible.

Also of note is the addition of metadata to zone catalog entries. This
can be employed by library users however they'd like, and it will also
eventually allow users to pass per-zone configuration information (like
zone-transfer ACLs) to the Server.

There is still significant work to be done. Perhaps most notably, the
zone API will have to be revised down the line to allow dynamic updates
and to be transactional (so that all lookups involved in a query see a
consistent snapshot of the zone). However, this commits gets the ball
rolling.