Rewrite parse_compressed_name without NameBuilder
Flamegraphs showed that parse_compressed_name was spending a large
amount of time in memcpy. Looking at the assembly that LLVM generates, I
discovered that the entire NameBuilder structure (approaching 400 bytes)
was being memcpy'd from one part of the stack frame to another. This was
due to NameBuilder::finish receiving self by value. For some reason,
it's not being optimized away as expected (possibly related to Rust's
issue 42763).

This replaces parse_compressed_name with a version that does not use
NameBuilder. This eliminates the memcpy and reduces the stack frame
size. Since this function is generally called at least once for every
message received, it seemed worth it. (For instance, consider that for
every EDNS message, this is called to parse the OPT record's owner,
which should always be the one-octet-long root name. Much more memory
was being copied when the NameBuilder structure was moved within the
stack frame than in the reconstruction of the domain name itself!)