Add basic owner domain name compression in Writer
This introduces compression to the owner field for resource records.
The compression works on the basis of "compression hints" that indicate
where a domain name previously has been written in the message. The
current hints available are "QNAME" and "most recent owner." The idea to
use such hints is inspired by Knot DNS, though I haven't actually
studied how its compression implementation works in much detail.

This does not produce complete/optimal compression in all cases. Indeed,
it does not perform any compression on domain names embedded in RDATA.
However, it does significantly decrease message size with very little
overhead, and in some of the simplest and most common cases (e.g.
answers for A and AAAA records, which include only a QNAME), the
compression is actually optimal. Furthermore, by operating so quickly in
common cases, it should be a good "frontend" to any full compression
implementation added later.