Implement AsRef, not Borrow, for RDATA types
Borrow is really meant for different kinds of data storage/ownership/
borrowing. If U implements Borrow<T>, it's required that Eq, Ord, and
Hash are equivalent for U and T.

Rdata is not an owned form of [u8], and even though we don't implement
the aforementioned traits for Rdata, it's conceptually not true that Eq,
Ord, and Hash are the same for Rdata and [u8]. Rdata equality is not
bitwise equality, for instance. Therefore, it seems like implementing
Borrow<[u8]> for Rdata was a mistake. Though CharacterString is closer
to plain-old [u8] than Rdata, I'm still unsure, so this removes
Borrow<[u8]> for CharacterString too.

The Borrow<[u8]> implementations have been convered to AsRef<[u8]> for
both types. AsRef is more general and the intent (cheap conversion)
definitely applies in both cases.