Use control messages only for 0.0.0.0 and ::
Control messages used to get the local destination address (recvmsg) and
set the local source address (sendmsg) are needed only when a socket is
bound to an unspecified address (0.0.0.0 for IPv4 and :: for IPv6). On
the platforms for which we've implemented local address support (Linux
and NetBSD), it doesn't hurt to use such control messages even with
sockets that are bound to a specific address. However, on FreeBSD, the
IP_SENDSRCADDR control message we'll need to use with sendmsg fails when
a specific IPv4 address is provided and the socket is bound to a
specific IPv4 address, even when those addresses are the same.
Therefore, in preparation for FreeBSD support, use control messages only
for sockets bound to unspecified addresses.