Support local address selection on FreeBSD
This uses the IP_RECVDSTADDR and IP_SENDSRCADDR control messages (which
are actually the same exact thing) for IPv4. Version 0.25 of the nix
crate is required for IP_SENDSRCADDR. For IPv6, the usual (RFC 3542)
IPV6_PKTINFO control message is used.

OpenBSD and DragonFly BSD also use the same control messages for this,
but since the nix crate currently does not have full support for
IPV6_PKTINFO, we'll have to add support for the two later. NetBSD also
has IP_RECVDSTADDR and IP_SENDSRCADDR, but we're already using its other
(IP_PKTINFO) API for IPv4.