CLI: use only the file name to get the zone name
We were accidentally using the whole path (stripped of the .zone suffix)
to determine the zone name when a zone file was specified on the command
line. For example, passing "path/to/quandary.test.zone" would
incorrectly set the zone name to "path/to/quandary.test.". This corrects
the bug; the previous path now (correctly) gives the zone name
"quandary.test.".