Add rather than subtract in Writer::set_edns
The available field should always be at least the DNS header size, which
is greater than OPT_RECORD_SIZE, so the subtraction should not cause
integer overflow. But to make the code match the parallel TSIG method
about to go in in beneath it (where this is an actual concern), let's
add instead.