Give Server access to TSIG keys
This is presently implemented in the simplest way: a HashMap mapping key
names to their algorithms and secrets. In the future, it may be better
to create a TsigKeyStore trait (or something like that) to abstract away
the key storage method, but this gets us up and running for now.