Add TSIG support to the daemon
See the updated README.md for the configuration file format for TSIG
keys. Once a key is configured, clients may use it to secure any DNS
transaction with the server, including regular queries. Like zones, keys
are reloaded from the configuration file when the server receives
SIGHUP.

I'm not very happy with the representation of TSIG keys in the TOML
configuration, and how it's not consistent with zones. (TSIG keys are a
table, but zones are an array, since a zone table's key would have to be
a combination of the zone name and the zone class.) We might have to
switch to a custom configuration file format at some point. But for now,
this works.