Add the length octet when hashing Labels
This follows the behavior of the standard library, which ensures that
the hash of several slices or strings added to a hasher consecutively is
not the same as the hash of their concatenation. In our case, domain
names like "quandary.test." and "quandarytest." were generating the same
hash. Adding the length to the hash of each Label is a standard way to
avoid this.