Update information about the glue Internet Draft
The latest draft is now draft-ietf-dnsop-glue-is-not-optional-07. Since
draft 4, there has been a change of title and some changes in
terminology, which are reflected in the rephrasing of the
src/server/query.rs comment in this commit. However, the behavior
specified for handing glue for sibling domain name servers has not
changed, and our implementation is still in compliance with it should it
become an RFC.