Add Template structure
The new Template structure stores the state of a Writer, including a
copy of the message that it has written so far. It can then be used to
construct new Writers that start out with the same state. This is
designed with AXFR in mind, but it might be useful in any situation
requiring the serialization of many similar messages.