Remove "Shutdown complete" message
When a very large number of zones are loaded, it can take a noticeable
amount of time for all the data to be dropped at the end of try_running.
This message appears *before* the data is dropped, so it can be
confusing when the server does not exit immediately. Furthermore, it's
not necessary, since we already have an "Exiting with success" message
printed when the server actually does exit.