Replace MostRecentRdata with MostRecentNameInRdata
This isn't a functionally equivalent replacement. The old
MostRecentRdata hint would produce a pointer to the most recently added
RDATA in a message, regardless of whether that RDATA actually started
with a domain name. By contrast, MostRecentNameInRdata produces a
pointer to the most recent name that appeared in RDATA of a known type.
Notably, this isn't necessarily in the most recent RDATA in the message.
However, for the one current use of MostRecentRdata in the codebase,
namely CNAME handling, MostRecentNameInRdata *does* serve as a drop-in
replacement.

The new hint type has the nice property that it can't be used to create
pointers to non-domain-name data. However, the real reason for it is
that keeping track of the most recently written name in RDATA will be
very helpful for implementing heuristic-based compression for unhinted
names. It'd be nice to use the same plumbing for both tasks!