Add heuristic-based compression for unhinted names
Following Knot's strategy, we guess that a domain name in a message is
probably similar to the ones that come right before it. In our
implementation, we select no more than two prior domain names to compare
to the name being compressed. These are

1. the most recently written RR owner, or the QNAME as a fallback; and
2. the name most recently written in RDATA.

While this does not guarantee optimal compression, for an authoritative
server, it does produce optimal or near-optimal compression in many
cases.

An explanatory comment in the source provides the details of the
algorithm and the reasoning behind the choices made.