Use an AXFR request and response in TSIG tests
When I added TSIG support for subsequent messages, I hacked together
tests for it by changing the MAC in the response message already used
for testing. However, it would be better to use an actual signed
subsequent message generated by another implementation to check our own
work. To this end, this commit replaces the three messages previously
used for TSIG testing with an AXFR request and the first two messages of
the corresponding AXFR response, as generated with BIND 9's dig and
named.

(The blocker for this was getting unhinted compression working, so that
Writer would generate messages identical to BIND's. I couldn't work
around this, since AFAIK, BIND provides no way to turn off compression
in zone transfers.)