Blocking I/O: use only one TCP listener thread
On Unix systems, we use poll and accept calls in sequence to implement
"accept with a timeout." As a result of the poll system call, the kernel
wakes up *all* TCP listener threads when a single new connection becomes
available. This is the thundering herd problem.

To address this, we could remove the timeouts or (when possible) use
even more platform-specific code. However, the former is important for
graceful shutdown, and the latter would be too much for the blocking I/O
provider, which tries to remain mostly cross-platform. (The fact that
there's no good cross-platform way to accept a connection with a timeout
is frustrating enough!) Furthermore, only a very busy server will need
more than one thread devoted to accepting connections. Such a server
would almost certainly use a platform-specific I/O provider anyway (once
we have some!).

Therefore, it seems best simply to remove the ability to use more than
one TCP listener thread with the blocking I/O provider.