Improve handling of address record types
The Internet (IN) class has A and AAAA records, and the Chaosnet (CH)
class has A records. What address records (if any) to look for in any
other class is unclear, especially since HS is the only other class
currently defined. Therefore, this commit disables additional section
processing and address-related zone validation checks for classes that
aren't IN or CH. It also limits these operations to considering A
records when dealing with CH.

This includes a modification to the Zone trait's lookup_addrs method. We
now expect implementations (1) to return A records regardless of class,
and (2) to return AAAA records as well when the class is IN. It seems
unlikely that a new DNS class will be defined in the future, but if one
is, then giving it a new class-specific definition of the A record type
seems like the originally intended way (if not necessarily the practical
way) of introducing an address record type for it. Thus, always
returning A records (and letting the caller decide whether using them
makes sense) seems like a fair choice. (The truth is, however, that the
utter dominance of the IN class makes it hard to reason about non-IN
classes!)

The Addresses structure returned by lookup_addrs still has fixed A and
AAAA fields. I admit that this does seem kludgy. However, given that
we're virtually always working with the IN class, and seeing that CH can
piggyback on the structure by setting only the A field, I don't think
that this is worth changing at this time.

This commit doesn't address the Hesiod (HS) class. BIND recognizes
IN-style IPv4 A records in it (though not AAAA records), so perhaps
additional section processing and the relevant zone validation checks
should be re-enabled for it. This is far from clear, however, especially
since documentation on Hesiod is sparse. Furthermore, some sources seem
to indicate that using Hesiod with the IN class is now preferred. If
that's the case, then all this might be moot. For now, we won't take any
further action for HS.

It's worth noting that we still treat A records as the same across all
classes, which is wrong: RFC 1034 indicates a different RDATA format for
A records in CH. This should be addressed very soon.