Handle class-specific RR types correctly
DNS record types may be class-specific. The A record type, for example,
is specified both for the Internet (IN) and Chaosnet (CH) classes, with
different RDATA formats in each class. In a similar way, the AAAA, WKS,
and SRV record types are all defined specifically for the IN class. When
dealing with a record type with class-specific RDATA formats, we must
process RDATA in accordance with the class. E.g., equality and
validation of A records in IN is different from equality and validation
of A records in CH. Furthermore, when we see a class/type combination
that we don't recognize, we must consider the RDATA format to be
unknown, even if we do recognize the type as a class-specific type
with RDATA formats defined for other classes. (See RFC 3597 section 2.)
So, for example, we know how to process NS RDATA in any class, since NS
RDATA is defined the same way for all classes. However, we don't
recognize AAAA RDATA in the CH class, since we only know how AAAA RDATA
is defined in the IN class. Therefore, if we see an AAAA record in CH,
we must treat it as unknown in accordance with RFC 3597.

However, Quandary currently assumes that each record type is defined the
same way for all classes. For instance, when shown an A record in CH or
CLASS10, Quandary will expect it to carry an IPv4 address and will
perform comparisons and validations on this basis. This is not compliant
behavior.

This commit implements the correct behavior. This includes updating the
RDATA-handling APIs to take a class parameter alongside a type parameter
when necessary. I decided to place the class parameter before the type
parameter, since it mirrors the order in which these fields appear in
zone files. (Additionally, this is the order that dnspython uses!)