Use inlined format arguments
In Rust 1.67, the clippy::uninlined_format_args lint is enabled. By
default, it suggests that format-string arguments be inlined when all
such arguments can be. The Rust core itself has a commit addressing this
new lint, so it looks like this is the direction that the Rust default
style is going.

The lint does not examine uses of format strings in the anyhow or log
APIs, but this commit also applies the same style in those cases for
consistency.