Address clippy::unchecked_duration_subtraction
This lint was turned on by default in Rust 1.67 (though it's already
back in the "pedantic" group in Clippy's master branch). To appease
Clippy in CI, we'll use checked_sub and expect. (Currently, this is all
that Instant's Sub<Duration> implementation does under the hood!)

Regarding the expect message: the new, computed last refill time should
be between the original last refill time and now, both of which Instant
can represent. Thus, Instant should be able to represent the new last
refill time. If it can't, then there's almost certainly a bug somewhere.