CI: replace/remove --all flag for Clippy/cargo fmt
For Clippy, the --all flag is shared with `cargo check` and is actually
a deprecated synonym for --workspace. For cargo fmt, the --all flag has
a few uses related to workspaces and local/path-based dependencies.
We're not using workspaces or local dependencies, so for us --all
doesn't make sense in either case.

In Clippy's case, the --all-targets flag does what we actually want
here, namely, to check/lint tests, examples, and benches in addition to
the default targets.