Add a Tokio I/O provider
The new Tokio I/O provider uses Rust's async features and Tokio to
communicate over the network and to achieve concurrency. The user
provides the Tokio runtime; the provider will work on both the
multi-thread and current-thread runtime flavors.

Unlike the blocking provider, which uses a receive/handle/send loop for
UDP, the Tokio I/O provider uses a receive loop that spawns a new Tokio
task to handle each received message and to send the response (if any).
Some admittedly informal tests showed that this has significantly better
performance than a receive/handle/send loop. Flamegraphs revealed that
with a receive/handle/send loop, a significant amount of time is spent
dealing with thread parking. I'm not an expert on Tokio internals, and I
didn't take a close look, but my guess is that this is related to
Tokio's strategy of using a single I/O driver per runtime, control of
which can be passed between threads.

On Linux, NetBSD, and FreeBSD, local address selection for UDP sockets
is supported, just as it is for blocking UDP sockets. This is
accomplished by reusing the existing custom syscall code in conjunction
with Tokio's AsyncFd type.

The Quandary daemon cannot yet use the Tokio I/O provider. Support in
the daemon is forthcoming.

To build/use the Tokio I/O provider, enable the new "tokio" feature.